refactor: divise to conquer
